requirements.txt
Django==1.8
django-crispy-forms==1.6.1
django-filter==1.0.4
django-suit==0.2.25
djangorestframework==3.6.3
futures==3.1.1
django-registration-redux==1.7
django-pipeline==1.6.13
